Refactored the Color Picker file to move out the landscape and portrait views
